add exit pwd and cd builtins, add some info to the prompt and also add some functions to str library

+++ src/util/string/string_util3.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   string_util3.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: edos-san <edos-san@student.42.fr>          +#+  +:+       +#+        */
+/*   By: afpachec <afpachec@42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/23 15:38:15 by edos-san          #+#    #+#             */
-/*   Updated: 2024/12/17 12:42:58 by edos-san         ###   ########.fr       */
+/*   Updated: 2025/01/09 02:03:53 by afpachec         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 #include <stdio.h>
 #include <ft_util.h>
 
-static char	*convert_base(char *s, int n, char *base, int size_base)
+static char	*convert_base(char *s, long long n, char *base, int size_base)
 {
 	if (n <= -10 || n >= 10)
 		s = convert_base(s, (n / size_base), base, size_base);
@@ -35,7 +35,7 @@ char	**__split_spacer(char const *str1, char *spacer)
 	return (__split(str1, spacer, 0, 0));
 }
 
-char	*__itoa(int n)
+char	*__itoa(long long n)
 {
 	char	buff[20];
